Update README, remove dead code, use env vars during SLOBS rebuild
* If you want to build mac-virtual-cam, you will need to set the OBS_CODESIGN_TEAM
in your build environment otherwise the obs cmake files will disable the feature.

+++ slobs-to-desktop.sh
@@ -5,24 +5,32 @@ rm -rf ./build_macos/packed_build/OBS.app
 rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/OBS.app
 
 echo "Time to run xcodebuild"
-#xcodebuild -project obs-studio.xcodeproj -target obs-studio -destination "generic/platform=macOS,name=Any Mac" -configuration RelWithDebInfo -parallelizeTargets -hideShellScriptEnvironment build
+
 cmake --build --target install --preset macos -v
-# now copy into desktop
-echo "Now copying libs"
-#cp  ./libobs-opengl/RelWithDebInfo/libobs-opengl.dylib ../../desktop/node_modules/obs-studio-node/Frameworks
 
-# cant copy libobs-- I think I need to copy ovber everything else first
-rm -rf ./../desktop/node_modules/obs-studio-node/Frameworks
-#cp  -R ./build_macos/UI/RelWithDebInfo/OBS.app/Contents/Frameworks ../desktop/node_modules/obs-studio-node
+exit_status=$?
 
-# copy into desktop
-cp  -R ./build_macos/packed_build/OBS.app/Contents/Frameworks ../desktop/node_modules/obs-studio-node
-# copy to OSN
-rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/Frameworks
-cp  -R ./build_macos/packed_build/OBS.app/Contents/Frameworks ../obs-studio-node/streamlabs-build.app/libobs-src/
-cp -R ./build_macos/packed_build/OBS.app ../obs-studio-node/streamlabs-build.app/libobs-src
+if [ $exit_status -eq 0 ]; then
+    # now copy into desktop
+    echo "Now copying libs"
+    #cp  ./libobs-opengl/RelWithDebInfo/libobs-opengl.dylib ../../desktop/node_modules/obs-studio-node/Frameworks
 
-rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/include
-rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/lib
-cp  -R ./build_macos/packed_build/include ../obs-studio-node/streamlabs-build.app/libobs-src/
-cp  -R ./build_macos/packed_build/lib ../obs-studio-node/streamlabs-build.app/libobs-src/
+    # cant copy libobs-- I think I need to copy ovber everything else first
+    rm -rf ./../desktop/node_modules/obs-studio-node/Frameworks
+    #cp  -R ./build_macos/UI/RelWithDebInfo/OBS.app/Contents/Frameworks ../desktop/node_modules/obs-studio-node
+
+    # copy into desktop
+    cp  -R ./build_macos/packed_build/OBS.app/Contents/Frameworks ../desktop/node_modules/obs-studio-node
+    # copy to OSN
+    rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/Frameworks
+    cp  -R ./build_macos/packed_build/OBS.app/Contents/Frameworks ../obs-studio-node/streamlabs-build.app/libobs-src/
+    cp -R ./build_macos/packed_build/OBS.app ../obs-studio-node/streamlabs-build.app/libobs-src
+
+    rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/include
+    rm -rf ../obs-studio-node/streamlabs-build.app/libobs-src/lib
+    cp  -R ./build_macos/packed_build/include ../obs-studio-node/streamlabs-build.app/libobs-src/
+    cp  -R ./build_macos/packed_build/lib ../obs-studio-node/streamlabs-build.app/libobs-src/
+else
+  echo "failed running the install target with exit code $exit_status."
+fi
+
